added audio files (unmodified), fixed issues in code, updated the readme
