+++ chariot-client/audio/ambient/README.md
@@ -1,0 +1,27 @@
+# Ambient Credits
+
+**Space**
+
+[Space_01](https://freesound.org/people/nathanshadow/sounds/22453/) *sampling+*
+
+**Deep Ocean**
+
+[Deep_Ocean_01](https://freesound.org/people/be-steele/sounds/150286/) *attribution 3.0*
+
+**Beach**
+
+[Beach_01](https://freesound.org/people/Noted451/sounds/531015/) *cc*
+
+[Beach_02](https://freesound.org/people/juskiddink/sounds/98479/) *attribution 4.0*
+
+**Forest**
+
+[Forest_01](https://freesound.org/people/reinsamba/sounds/18765/) *attribution 4.0*
+
+[Forest_02](https://freesound.org/people/dobroide/sounds/22384/) *attribution 4.0*
+
+**City**
+
+[City_01](https://freesound.org/people/Ajexk/sounds/246171/) *cc*
+
+[City_02](https://freesound.org/people/sagetyrtle/sounds/36734/) *cc*
