+++ chariot-client/audio/sfx/README.md
@@ -1,0 +1,57 @@
+# SFX Credits
+
+**Rocket Thrusters:**
+
+[Thrust_01](https://freesound.org/people/Zovex/sounds/237974/) *cc*
+
+**Fast Turn:**
+
+[Turn_01](https://freesound.org/people/RutgerMuller/sounds/104026/) *cc*
+
+**Brake:**
+
+[Brake_01](https://freesound.org/people/Medartimus/sounds/217542/) *attribution*
+
+[Brake_02](https://freesound.org/people/brunoboselli/sounds/457294/) *cc*
+
+**Bump**
+
+[Bump_01](https://freesound.org/people/Kane53126/sounds/257928/) *cc*
+
+[Bump_02](https://freesound.org/people/Macif/sounds/207824/) *cc*
+
+**Crash**
+
+[Crash_01](https://freesound.org/people/pauliep83/sounds/33675/) *sampling+*
+
+[Crash_02](https://freesound.org/people/squareal/sounds/237375/) *cc*
+
+[Crash_03](https://freesound.org/people/Rock%20Savage/sounds/59263/) *sampling+*
+
+[Crash_04](https://freesound.org/people/sandyrb/sounds/95078/) *attribution 4.0*
+
+[Crash_05](https://freesound.org/people/CGEffex/sounds/99960/) *attribution 3.0*
+
+**Power Up**
+
+[Power_Get_01](https://freesound.org/people/TreasureSounds/sounds/332629/) *attribution 3.0*
+
+[Success_01](https://freesound.org/people/GabrielAraujo/sounds/242501/) *cc*
+
+[Effect_01](https://freesound.org/people/Breviceps/sounds/453388/) *cc*
+
+[Effect_02](https://freesound.org/people/suntemple/sounds/253172/) *cc*
+
+[Speed_Up_01](https://freesound.org/people/original_sound/sounds/376817/) *attribution 3.0*
+
+[Speed_Up_02](https://freesound.org/people/Eponn/sounds/420998/) *cc*
+
+[Speed_Down_01](https://freesound.org/people/Distanttone/sounds/231278/) *cc*
+
+[Throw_01](https://freesound.org/people/Pogotron/sounds/60835/) *sampling+*
+
+[Throw_02](https://freesound.org/people/copyc4t/sounds/214898/) *attribution 4.0*
+
+[Slip_01](https://freesound.org/people/fordps3/sounds/533233/)*attribution 3.0*
+
+[Slip_02](https://freesound.org/people/WannyManny/sounds/609936/) *cc*
